Updated the patches in the included directory...READ ON:
ATTENTION: At the time of writing, the organized edge detection stuff has just gone templated as of r6666.  This commit ports my code on top of the new version, and the included directory contains patches for backporting r6666 or r6701 of features/.../organized_edge_detection.* and [r6598,r6679] of the 2d subsystem onto PCL release 1.6.0.  Notice how I said the minimum version of 2d for use with the templated code is r6598; this is because this subsystem's API changed appreciably just before this revision was committed.  Unfortunately, the changes brought significant slowdowns---at least on my setup---as follows:
- r6598 takes about three times as long to perform the edge detection as before (at ~1.5s)
- r6601 is the "best" version to use, but has about the same speed cost as the above
- r6598 takes about sixty times as long to do the detection (at ~30s)
Because these times are rather prohibitive, I have ceased further efforts to use the latest code at this time.  However, I can only assume that the decided slowdown is either a bug that will soon be fixed or a silly mistake on my part that you'll easily solve.  This version of the code is still included in order to make it more future-proof for such time as the issues can be solved, but the last version I actually worked with and supported was that which used the pre-templated version of the edge detection.

Signing off,
Sol