Copied the high curvature edge detection code inline
In order to debug a major slowdown introduced by the templated version.
NOTE: This code will not compile unless the 2d subsystem's CMakeLists.txt contains instructions to compile in its HPP files.
NOTE: This code WILL segfault; that's not the point, as it exists in order to see where the delay occurs.